+++ BloomService.Web/Models/PermissionCodeModel.cs
@@ -2,7 +2,7 @@
 {
     public class PermissionCodeModel : EntityModel
     {
-        public int PERMISSIONCODE { get; set; }
+        public long PERMISSIONCODE { get; set; }
         public string QINACTIVE { get; set; }
         public string DESCRIPTION { get; set; }
     }
