+++ BloomService.Web/Models/RateSheetModel.cs
@@ -2,7 +2,7 @@
 {
     public class RateSheetModel : EntityModel
     {
-        public int RATESHEETNBR { get; set; }
+        public long RATESHEETNBR { get; set; }
         public string DESCRIPTION { get; set; }
         public double LABORRATE { get; set; }
         public double STOCKPARTSMARKUP { get; set; }
